--- src/main/java/com/arroyo/carlos/juan/repository/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.arroyo.carlos.juan.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.arroyo.carlos.juan.model.Customer;
-
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-
-}
